+++ kernel_tuner/core.py
@@ -310,6 +310,9 @@ class DeviceInterface(object):
                 return None
 
             #add constant memory arguments to compiled module
+            if kernel_options.smem_args is not None:
+                self.dev.copy_shared_memory_args(kernel_options.smem_args)
+            #add constant memory arguments to compiled module
             if kernel_options.cmem_args is not None:
                 self.dev.copy_constant_memory_args(kernel_options.cmem_args)
             #add texture memory arguments to compiled module