+++ test/test_cuda_functions.py
@@ -69,7 +69,7 @@ def test_compile():
         print(str(e))
         assert False
 
-def dummy_func(a, b, block=0, grid=0, stream=None, texrefs=None):
+def dummy_func(a, b, block=0, grid=0, stream=None, shared=0, texrefs=None):
     pass
 
 @skip_if_no_cuda