+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+
+### Added
+- kernelbuilder functionality for including kernels in Python applications
+- smem_args option for dynamically allocated shared memory in CUDA kernels
+
 ## [0.3.0] - 2019-12-20
 ### Changed
 - fix for output checking, custom verify functions are called just once